+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `tnt_cartridge_config_checksum` metric.
+
 ### Changed
 
 ### Fixed